requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-ppc64le
_libgcc_mutex=0.1=main
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.8.1=py39h140841e_1
aiosignal=1.2.0=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h140841e_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h6ffa863_0
blas=1.0=openblas
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h140841e_1003
bzip2=1.0.8=h6eb9509_2
c-ares=1.18.1=h140841e_0
ca-certificates=2022.3.29=h6ffa863_1
cachetools=5.0.0=pyhd8ed1ab_0
cartopy=0.18.0=py39h2fb373a_1
certifi=2021.10.8=py39h6ffa863_2
cffi=1.15.0=py39he068cbb_1
cftime=1.5.1.1=py39hfc2be9b_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=7.1.2=pyh9f0ad1d_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=35.0.0=py39h1b86913_0
curl=7.82.0=h140841e_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py39h29c3540_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=py_0
entrypoints=0.4=py39h6ffa863_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h9215f1b_0
frozenlist=1.2.0=py39h140841e_0
future=0.18.2=py39h6ffa863_1
geos=3.8.0=he6710b0_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
google-auth=2.6.5=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.42.0=py39hedb86c2_0
hdf4=4.2.13=h2ab409d_3
hdf5=1.10.6=hb1b8bf9_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.3=py39hc1b9086_1
ipykernel=6.9.1=py39h6ffa863_0
ipython=8.2.0=py39h6ffa863_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39h6ffa863_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9d=h140841e_0
jsonschema=4.4.0=py39h6ffa863_0
jupyter=1.0.0=py39h6ffa863_7
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.9.2=py39h6ffa863_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.2=py39h29c3540_0
krb5=1.19.2=h6205695_0
lcms2=2.12=h2045e0b_0
ld_impl_linux-ppc64le=2.33.1=h0f24833_7
libcurl=7.82.0=ha47bf17_0
libedit=3.1.20210910=h140841e_0
libev=4.33=h140841e_1
libffi=3.3=he6710b0_2
libgcc-ng=8.2.0=h822a55f_1
libgfortran-ng=7.3.0=h822a55f_1
libnetcdf=4.8.1=h37953bf_1
libnghttp2=1.46.0=hedb86c2_0
libopenblas=0.3.18=h04d2106_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h727906f_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=8.2.0=h822a55f_1
libtiff=4.2.0=h781710b_0
libwebp=1.2.2=he32dc1f_0
libwebp-base=1.2.2=h140841e_0
libzip=1.5.1=h2a439d1_1003
lz4-c=1.9.3=h29c3540_1
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.0.1=py39h140841e_0
matplotlib=3.5.1=py39h6ffa863_1
matplotlib-base=3.5.1=py39h29c3540_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h140841e_1000
multidict=5.2.0=py39h140841e_2
munkres=1.1.4=py_0
nbclient=0.5.13=py39h6ffa863_0
nbconvert=6.4.4=py39h6ffa863_0
nbformat=5.3.0=py39h6ffa863_0
ncurses=6.3=h140841e_2
nest-asyncio=1.5.5=py39h6ffa863_0
netcdf4=1.5.7=py39hf727584_1
ninja=1.10.2=py39hde33034_3
notebook=6.4.8=py39h6ffa863_0
numpy=1.21.5=py39h76732f4_1
numpy-base=1.21.5=py39h37faa36_1
oauthlib=3.2.0=pyhd8ed1ab_0
openssl=1.1.1n=h140841e_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=py_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39hff93520_0
pip=21.2.4=py39h6ffa863_0
proj=7.0.1=h59a7b90_1
prometheus_client=0.13.1=pyhd3eb1b0_0
promise=2.3=py39hc1b9086_5
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.1=py39h29c3540_0
psutil=5.8.0=py39h140841e_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py39h7a5889e_0
pyshp=2.1.3=pyhd3eb1b0_0
pysocks=1.7.1=py39hc1b9086_5
python=3.9.12=h836d2c2_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.10.2=cpu_py39h6f0ae12_0
pytorch-lightning=0.8.5=py_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=py39h140841e_1
pyzmq=22.3.0=py39h29c3540_2
readline=8.1.2=h140841e_1
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scipy=1.7.3=py39he743248_0
send2trash=1.8.0=pyhd3eb1b0_1
sentry-sdk=1.5.10=pyhd8ed1ab_0
setproctitle=1.2.2=py39h140841e_1004
setuptools=61.2.0=py39h6ffa863_0
shapely=1.7.1=py39h26044be_0
shortuuid=1.0.8=py39hc1b9086_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.2=hd7247d8_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.8.0=pyhd8ed1ab_1
tensorboard-data-server=0.6.0=py39h8068876_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
terminado=0.13.1=py39h6ffa863_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.11=h7e00dab_0
torchvision=0.2.2=py_3
tornado=6.1=py39h140841e_0
tqdm=4.63.0=pyhd3eb1b0_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=pyhd8ed1ab_0
wandb=0.12.14=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h6ffa863_1
werkzeug=2.1.1=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39h6ffa863_0
xz=5.2.5=h140841e_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py39h140841e_0
zeromq=4.3.4=h29c3540_0
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h140841e_1
zstd=1.4.9=hc52992f_0
